Add firestore uploading for sections only
