+++ internal/firebase/courses.go
@@ -4,132 +4,11 @@ import (
 	"context"
 	"fmt"
 	"strings"
-	"time"
 
 	"cloud.google.com/go/firestore"
 	"github.com/acmutd/acmutd-api/internal/types"
 	"google.golang.org/api/iterator"
 )
-
-func ensureSectionDocID(course types.Course, term string) string {
-	if section := sanitizeDocID(course.SectionAddress); section != "" {
-		return strings.ToLower(section)
-	}
-
-	prefix := sanitizeDocID(strings.ToLower(strings.TrimSpace(course.CoursePrefix)))
-	number := sanitizeDocID(strings.ToLower(strings.TrimSpace(course.CourseNumber)))
-	section := sanitizeDocID(strings.ToLower(course.Section))
-	if section == "" {
-		section = "000"
-	}
-	normalizedTerm := sanitizeDocID(strings.ToLower(strings.TrimSpace(term)))
-
-	generated := fmt.Sprintf("%s%s.%s.%s", prefix, number, section, normalizedTerm)
-	generated = strings.ReplaceAll(generated, "..", ".")
-	generated = strings.Trim(generated, ".")
-
-	return generated
-}
-
-type preparedCourse struct {
-	Course    types.Course
-	PrefixID  string
-	NumberID  string
-	SectionID string
-}
-
-func prepareCourseForTerm(course types.Course, term string) (preparedCourse, bool) {
-
-	normalizedTerm := strings.ToLower(strings.TrimSpace(term))
-
-	if normalizedTerm == "" {
-		return preparedCourse{}, false
-	}
-
-	course.Term = normalizedTerm
-	course.CoursePrefix = strings.ToLower(strings.TrimSpace(course.CoursePrefix))
-	course.CourseNumber = strings.ToLower(strings.TrimSpace(course.CourseNumber))
-	course.Section = strings.ToLower(strings.TrimSpace(course.Section))
-
-	prefixID := sanitizeDocID(course.CoursePrefix)
-	numberID := sanitizeDocID(course.CourseNumber)
-	if prefixID == "" || numberID == "" {
-		return preparedCourse{}, false
-	}
-
-	sectionID := ensureSectionDocID(course, normalizedTerm)
-	course.SectionAddress = sectionID
-
-	return preparedCourse{
-		Course:    course,
-		PrefixID:  prefixID,
-		NumberID:  numberID,
-		SectionID: sectionID,
-	}, true
-}
-
-func (c *Firestore) sectionsCollection(prefixID, numberID string) *firestore.CollectionRef {
-	return c.Collection("courses").
-		Doc(prefixID).
-		Collection("numbers").
-		Doc(numberID).
-		Collection("sections")
-}
-
-/*
-Structure:
-
-  - courses/{course_prefix}/numbers/{course_number}/sections/{section_address}
-
-  - terms/{term}/prefixes/{course_prefix}
-
-    This mirrors the recommended Firestore layout for efficient collection group
-    queries by term, course prefix, and course number while maintaining fast prefix
-    lookups for each term.
-*/
-func (c *Firestore) InsertClassesWithIndexes(ctx context.Context, courses []types.Course, term string) {
-	writer := c.BulkWriter(ctx)
-	defer writer.End()
-
-	normalizedTerm := strings.ToLower(strings.TrimSpace(term))
-	if normalizedTerm == "" {
-		return
-	}
-
-	termDoc := c.Collection("terms").Doc(normalizedTerm)
-	writer.Set(termDoc, map[string]any{
-		"term":         normalizedTerm,
-		"last_updated": time.Now(),
-	}, firestore.MergeAll)
-
-	prefixes := make(map[string]string)
-
-	for _, course := range courses {
-		prepared, ok := prepareCourseForTerm(course, normalizedTerm)
-		if !ok {
-			continue
-		}
-
-		doc := c.sectionsCollection(prepared.PrefixID, prepared.NumberID).Doc(prepared.SectionID)
-		writer.Set(doc, prepared.Course)
-
-		if _, exists := prefixes[prepared.PrefixID]; !exists {
-			prefixes[prepared.PrefixID] = prepared.Course.CoursePrefix
-		}
-	}
-
-	for prefixID, originalPrefix := range prefixes {
-		writer.Set(
-			termDoc.Collection("prefixes").Doc(prefixID),
-			map[string]any{
-				"course_prefix":     originalPrefix,
-				"normalized_prefix": prefixID,
-				"term":              normalizedTerm,
-			},
-			firestore.MergeAll,
-		)
-	}
-}
 
 func (c *Firestore) QueryCourses(ctx context.Context, q types.CourseQuery) ([]types.Course, bool, error) {
 	term := strings.ToLower(strings.TrimSpace(q.Term))
