+++ internal/uploader/uploader.go
@@ -2,6 +2,7 @@ package uploader
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"log"
 	"os"
@@ -14,8 +15,8 @@ import (
 var folders = []string{"coursebook", "enhanced_grades", "professors"}
 
 type UploaderHandler struct {
-	client *firebase.FBClient
-	config UploaderConfig
+	fbclient *firebase.FBClient
+	config   UploaderConfig
 }
 
 type UploaderConfig struct {
@@ -34,7 +35,7 @@ func NewUploaderHandler(client *firebase.FBClient, saveEnv string, gather bool) 
 	}
 
 	handler := &UploaderHandler{
-		client: client,
+		fbclient: client,
 		config: UploaderConfig{
 			SaveEnvironment: saveEnv,
 			ShouldGather:    gather,
@@ -59,7 +60,7 @@ func parseClassTerms(raw string) []string {
 func (h *UploaderHandler) Start() error {
 	if h.config.ShouldGather {
 		log.Printf("Gathering data for terms: %v", h.config.ClassTerms)
-		if err := h.client.CloudStorage().DownloadFolders(context.Background(), h.config.InputBase, folders); err != nil {
+		if err := h.fbclient.CloudStorage().DownloadFolders(context.Background(), h.config.InputBase, folders); err != nil {
 			return fmt.Errorf("gather phase failed: %w", err)
 		}
 		log.Println("All data gathered successfully")
@@ -74,6 +75,13 @@ func (h *UploaderHandler) Start() error {
 
 	log.Printf("Combined %d section documents and %d course groups across terms %v", len(sections), len(courses), h.config.ClassTerms)
 
-	// TODO: write sections to Firestore
+	if math, ok := courses["math"]; ok {
+		if c, ok := math["2415"]; ok {
+			data, _ := json.MarshalIndent(c, "", "  ")
+			log.Printf("courses[\"math\"][\"2415\"]:\n%s", string(data))
+		}
+	}
+
+	h.InsertClassesWithIndexes(context.Background(), sections, h.config.ClassTerms[0])
 	return nil
 }
